Reorganized code. Refined docs and README. Fixed #38 and #46. New code unit tested
