+++ docs/model_components.rst
@@ -5,9 +5,9 @@ This module contains the four main components that will comprise a Wide and
 Deep model, and the ``WideDeep`` "constructor" class. These four components
 are: ``wide``, ``deeptabular``, ``deeptext``, ``deepimage``.
 
-.. note:: ``TabMlp``, ``TabResnet``, ``TabNet`` and ``TabTransformer`` can all
-    be used as the ``deeptabular``  component of the model and simply
-    represent different alternatives
+.. note:: ``TabMlp``, ``TabResnet``, ``TabNet``, ``TabTransformer`` and ``SAINT`` can
+	all be used as the ``deeptabular``  component of the model and simply
+	represent different alternatives
 
 .. autoclass:: pytorch_widedeep.models.wide.Wide
 	:exclude-members: forward
@@ -25,7 +25,11 @@ are: ``wide``, ``deeptabular``, ``deeptext``, ``deepimage``.
 	:exclude-members: forward
 	:members:
 
-.. autoclass:: pytorch_widedeep.models.tab_transformer.TabTransformer
+.. autoclass:: pytorch_widedeep.models.transformers.tab_transformer.TabTransformer
+	:exclude-members: forward
+	:members:
+
+.. autoclass:: pytorch_widedeep.models.transformers.saint.SAINT
 	:exclude-members: forward
 	:members:
