+++ pytorch_widedeep/models/__init__.py
@@ -5,7 +5,5 @@ from pytorch_widedeep.models.wide_deep import WideDeep
 from pytorch_widedeep.models.deep_image import DeepImage
 from pytorch_widedeep.models.tab_resnet import TabResnet
 from pytorch_widedeep.models.tabnet.tab_net import TabNet
-from pytorch_widedeep.models.transformers.transformer_models import (
-    SAINT,
-    TabTransformer,
-)
+from pytorch_widedeep.models.transformers.saint import SAINT
+from pytorch_widedeep.models.transformers.tab_transformer import TabTransformer
