+++ pytorch_widedeep/models/tab_mlp.py
@@ -34,6 +34,11 @@ def dense_layer(
     linear_first: bool,
 ):
     # This is basically the LinBnDrop class at the fastai library
+    if activation == "geglu":
+        raise ValueError(
+            "'geglu' activation is only used as 'transformer_activation' "
+            "in transformer-based models (TabTransformer and SAINT)"
+        )
     act_fn = _get_activation_fn(activation)
     layers = [nn.BatchNorm1d(out if linear_first else inp)] if bn else []
     if p != 0:
@@ -92,11 +97,21 @@ class TabMlp(nn.Module):
         Dict containing the index of the columns that will be passed through
         the TabMlp model. Required to slice the tensors. e.g. {'education':
         0, 'relationship': 1, 'workclass': 2, ...}
+    embed_input: List, Optional, default = None
+        List of Tuples with the column name, number of unique values and
+        embedding dimension. e.g. [(education, 11, 32), ...]
+    embed_dropout: float, default = 0.1
+        embeddings dropout
+    continuous_cols: List, Optional, default = None
+        List with the name of the numeric (aka continuous) columns
+    batchnorm_cont: bool, default = False
+        Boolean indicating whether or not to apply batch normalization to the
+        continuous input
     mlp_hidden_dims: List, default = [200, 100]
         List with the number of neurons per dense layer in the mlp.
     mlp_activation: str, default = "relu"
         Activation function for the dense layers of the MLP. Currently
-        only "relu", "leaky_relu" and "gelu" are supported
+        'relu', 'leaky_relu' and 'gelu' are supported
     mlp_dropout: float or List, default = 0.1
         float or List of floats with the dropout between the dense layers.
         e.g: [0.5,0.5]
@@ -110,16 +125,6 @@ class TabMlp(nn.Module):
         Boolean indicating the order of the operations in the dense
         layer. If ``True: [LIN -> ACT -> BN -> DP]``. If ``False: [BN -> DP ->
         LIN -> ACT]``
-    embed_input: List, Optional, default = None
-        List of Tuples with the column name, number of unique values and
-        embedding dimension. e.g. [(education, 11, 32), ...]
-    embed_dropout: float, default = 0.1
-        embeddings dropout
-    continuous_cols: List, Optional, default = None
-        List with the name of the numeric (aka continuous) columns
-    batchnorm_cont: bool, default = False
-        Boolean indicating whether or not to apply batch normalization to the
-        continuous input
 
     Attributes
     ----------
@@ -148,16 +153,16 @@ class TabMlp(nn.Module):
     def __init__(
         self,
         column_idx: Dict[str, int],
+        embed_input: Optional[List[Tuple[str, int, int]]] = None,
+        embed_dropout: float = 0.1,
+        continuous_cols: Optional[List[str]] = None,
+        batchnorm_cont: bool = False,
         mlp_hidden_dims: List[int] = [200, 100],
         mlp_activation: str = "relu",
         mlp_dropout: Union[float, List[float]] = 0.1,
         mlp_batchnorm: bool = False,
         mlp_batchnorm_last: bool = False,
         mlp_linear_first: bool = False,
-        embed_input: Optional[List[Tuple[str, int, int]]] = None,
-        embed_dropout: float = 0.1,
-        continuous_cols: Optional[List[str]] = None,
-        batchnorm_cont: bool = False,
     ):
         super(TabMlp, self).__init__()
