+++ pytorch_widedeep/models/transformers/tab_transformer.py
@@ -4,7 +4,6 @@ from torch import nn
 from pytorch_widedeep.wdtypes import *  # noqa: F403
 from pytorch_widedeep.models.tab_mlp import MLP
 from pytorch_widedeep.models.transformers.layers import (
-    SaintEncoder,
     SharedEmbeddings,
     TransformerEncoder,
     FullEmbeddingDropout,
@@ -13,7 +12,7 @@ from pytorch_widedeep.models.transformers.layers import (
 
 class TabTransformer(nn.Module):
     r"""Adaptation of TabTransformer model
-    (https://arxiv.org/pdf/2012.06678.pdf) model that can be used as the
+    (https://arxiv.org/abs/2012.06678) model that can be used as the
     deeptabular component of a Wide & Deep model.
 
     Parameters
@@ -25,14 +24,12 @@ class TabTransformer(nn.Module):
     embed_input: List
         List of Tuples with the column name and number of unique values
         e.g. [(education, 11), ...]
-    continuous_cols: List, Optional, default = None
-        List with the name of the numeric (aka continuous) columns
     embed_dropout: float, default = 0.1
         Dropout to be applied to the embeddings matrix
     full_embed_dropout: bool, default = False
-        Boolean indicating if an entire embedding (i.e. the representation
-        for one categorical column) will be dropped in the batch. See:
-        :obj:`pytorch_widedeep.model.tab_transformer.FullEmbeddingDropout`.
+        Boolean indicating if an entire embedding (i.e. the representation of
+        one column) will be dropped in the batch. See:
+        :obj:`pytorch_widedeep.models.transformers.layers.FullEmbeddingDropout`.
         If ``full_embed_dropout = True``, ``embed_dropout`` is ignored.
     shared_embed: bool, default = False
         The idea behind ``shared_embed`` is described in the Appendix A in the paper:
@@ -42,10 +39,21 @@ class TabTransformer(nn.Module):
     add_shared_embed: bool, default = False,
         The two embedding sharing strategies are: 1) add the shared embeddings to the column
         embeddings or 2) to replace the first ``frac_shared_embed`` with the shared
-        embeddings. See :obj:`pytorch_widedeep.models.tab_transformer.SharedEmbeddings`
-    frac_shared_embed: int, default = 8
+        embeddings. See :obj:`pytorch_widedeep.models.transformers.layers.SharedEmbeddings`
+    frac_shared_embed: float, default = 0.25
         The fraction of embeddings that will be shared by all the different categories for
         one particular column.
+    continuous_cols: List, Optional, default = None
+        List with the name of the numeric (aka continuous) columns
+    embed_continuous: bool, default = False,
+        Boolean indicating if the continuous features will be "embedded"
+        (i.e. each passed through a 1 layer MLP with Relu). Note that setting
+        this to true is equivalent to the so called `FT-Transformer
+        <https://arxiv.org/abs/2106.11959>`_ (Feature Tokenizer +
+        Transformer).
+    cont_norm_layer: str, default =  "layernorm",
+        Type of normalization layer applied to the continuous features if they
+        are not embedded. Options are: 'layernorm' or 'batchnorm'.
     input_dim: int, default = 32
         The so-called *dimension of the model*. Is the number of embeddings used to encode
         the categorical columns
@@ -54,32 +62,23 @@ class TabTransformer(nn.Module):
     n_blocks: int, default = 6
         Number of Transformer blocks
     dropout: float, default = 0.1
-        Dropout that will be applied internally to the
-        ``TransformerEncoder`` (see
-        :obj:`pytorch_widedeep.model.tab_transformer.TransformerEncoder`) and the
-        output MLP
+        Dropout that will be applied internally to the ``TransformerEncoder``
+        (see :obj:`pytorch_widedeep.models.transformers.layers.TransformerEncoder`)
+        and the output MLP
     keep_attn_weights: bool, default = False
         If set to ``True`` the model will store the attention weights in the ``attention_weights``
         attribute.
     ff_hidden_dim: int, default = 128
         Hidden dimension of the ``FeedForward`` Layer. See
-        :obj:`pytorch_widedeep.model.tab_transformer.FeedForward`.
+        :obj:`pytorch_widedeep.models.transformers.layers.FeedForward`.
     transformer_activation: str, default = "gelu"
-        Transformer Encoder activation function
-    with_special_token: bool, default = False,
-        Boolean indicating if the special token ``[CLS]`` was used during the
-        data pre-processing
-    embed_continuous: bool, default = False,
-        Boolean indicating if the continuous features will be "embedded"
-        (i.e. each passed through a 1 layer MLP)
-    cont_norm_layer: str, default =  "layernorm",
-        Type of normalization layer applied to the continuous features if they
-        are not embedded. Options are: 'layernorm' or 'batchnorm'.
+        Transformer Encoder activation function. 'relu', 'leaky_relu', 'gelu'
+        and 'geglu' are supported
     mlp_hidden_dims: List, Optional, default = None
         MLP hidden dimensions. If not provided it will default to ``[4*l,
         2*l]`` where ``l`` is the mlp input dimension
-    mlp_activation: str, default = "gelu"
-        MLP activation function
+    mlp_activation: str, default = "relu"
+        MLP activation function. 'relu', 'leaky_relu' and 'gelu' are supported
     mlp_batchnorm: bool, default = False
         Boolean indicating whether or not to apply batch normalization to the
         dense layers
@@ -102,7 +101,7 @@ class TabTransformer(nn.Module):
     transformer_blks: ``nn.Sequential``
         Sequence of Transformer blocks
     attention_weights: List
-        List with the attention weights per block
+        List with the attention weights per block if ``keep_attn_weights = True``.
     transformer_mlp: ``nn.Module``
         MLP component in the model
     output_dim: int
@@ -126,12 +125,14 @@ class TabTransformer(nn.Module):
         self,
         column_idx: Dict[str, int],
         embed_input: List[Tuple[str, int]],
-        continuous_cols: Optional[List[str]] = None,
         embed_dropout: float = 0.1,
         full_embed_dropout: bool = False,
         shared_embed: bool = False,
         add_shared_embed: bool = False,
-        frac_shared_embed: int = 8,
+        frac_shared_embed: float = 0.25,
+        continuous_cols: Optional[List[str]] = None,
+        embed_continuous: bool = False,
+        cont_norm_layer: str = "layernorm",
         input_dim: int = 32,
         n_heads: int = 8,
         n_blocks: int = 6,
@@ -139,9 +140,6 @@ class TabTransformer(nn.Module):
         keep_attn_weights: bool = False,
         ff_hidden_dim: int = 32 * 4,
         transformer_activation: str = "gelu",
-        with_special_token: bool = False,
-        embed_continuous: bool = False,
-        cont_norm_layer: str = "layernorm",
         mlp_hidden_dims: Optional[List[int]] = None,
         mlp_activation: str = "relu",
         mlp_batchnorm: bool = False,
@@ -152,12 +150,14 @@ class TabTransformer(nn.Module):
 
         self.column_idx = column_idx
         self.embed_input = embed_input
-        self.continuous_cols = continuous_cols
         self.embed_dropout = embed_dropout
         self.full_embed_dropout = full_embed_dropout
         self.shared_embed = shared_embed
         self.add_shared_embed = add_shared_embed
         self.frac_shared_embed = frac_shared_embed
+        self.continuous_cols = continuous_cols
+        self.embed_continuous = embed_continuous
+        self.cont_norm_layer = cont_norm_layer
         self.input_dim = input_dim
         self.n_heads = n_heads
         self.n_blocks = n_blocks
@@ -165,20 +165,13 @@ class TabTransformer(nn.Module):
         self.keep_attn_weights = keep_attn_weights
         self.ff_hidden_dim = ff_hidden_dim
         self.transformer_activation = transformer_activation
-        self.with_special_token = with_special_token
-        self.embed_continuous = embed_continuous
-        self.cont_norm_layer = cont_norm_layer
         self.mlp_hidden_dims = mlp_hidden_dims
         self.mlp_activation = mlp_activation
         self.mlp_batchnorm = mlp_batchnorm
         self.mlp_batchnorm_last = mlp_batchnorm_last
         self.mlp_linear_first = mlp_linear_first
 
-        if "special_token" in self.column_idx and not self.with_special_token:
-            raise ValueError(
-                "The data was pre-processed using the 'CLS' special token,"
-                " Please set 'with_special_token' to True"
-            )
+        self.with_cls_token = "cls_token" in self.column_idx
 
         self._set_categ_embeddings()
 
@@ -247,7 +240,7 @@ class TabTransformer(nn.Module):
                 else:
                     self.attention_weights[i] = blk.self_attn.attn_weights
 
-        if self.with_special_token:
+        if self.with_cls_token:
             x = x[:, 0, :]
         else:
             x = x.flatten(1)
@@ -308,7 +301,7 @@ class TabTransformer(nn.Module):
 
     def _set_mlp_hidden_dims(self) -> List[int]:
         if self.continuous_cols is not None:
-            if self.with_special_token:
+            if self.with_cls_token:
                 if self.embed_continuous:
                     mlp_hidden_dims = [
                         self.input_dim,
@@ -332,188 +325,3 @@ class TabTransformer(nn.Module):
             mlp_inp_l = len(self.embed_input) * self.input_dim
             mlp_hidden_dims = [mlp_inp_l, mlp_inp_l * 4, mlp_inp_l * 2]
         return mlp_hidden_dims
-
-
-class SAINT(TabTransformer):
-    r"""Adaptation of SAINT model
-    (https://arxiv.org/abs/2106.01342) model that can be used as the
-    deeptabular component of a Wide & Deep model.
-
-    Parameters for this model are identical to those of the ``TabTransformer``
-
-    Parameters
-    ----------
-    column_idx: Dict
-        Dict containing the index of the columns that will be passed through
-        the DeepDense model. Required to slice the tensors. e.g. {'education':
-        0, 'relationship': 1, 'workclass': 2, ...}
-    embed_input: List
-        List of Tuples with the column name and number of unique values
-        e.g. [(education, 11), ...]
-    continuous_cols: List, Optional, default = None
-        List with the name of the numeric (aka continuous) columns
-    embed_dropout: float, default = 0.1
-        Dropout to be applied to the embeddings matrix
-    full_embed_dropout: bool, default = False
-        Boolean indicating if an entire embedding (i.e. the representation
-        for one categorical column) will be dropped in the batch. See:
-        :obj:`pytorch_widedeep.model.tab_transformer.FullEmbeddingDropout`.
-        If ``full_embed_dropout = True``, ``embed_dropout`` is ignored.
-    shared_embed: bool, default = False
-        The idea behind ``shared_embed`` is described in the Appendix A in the paper:
-        `'The goal of having column embedding is to enable the model to distinguish the
-        classes in one column from those in the other columns'`. In other words, the idea
-        is to let the model learn which column is embedding at the time.
-    add_shared_embed: bool, default = False,
-        The two embedding sharing strategies are: 1) add the shared embeddings to the column
-        embeddings or 2) to replace the first ``frac_shared_embed`` with the shared
-        embeddings. See :obj:`pytorch_widedeep.models.tab_transformer.SharedEmbeddings`
-    frac_shared_embed: int, default = 8
-        The fraction of embeddings that will be shared by all the different categories for
-        one particular column.
-    input_dim: int, default = 32
-        The so-called *dimension of the model*. Is the number of embeddings used to encode
-        the categorical columns
-    n_heads: int, default = 8
-        Number of attention heads per Transformer block
-    n_blocks: int, default = 6
-        Number of Transformer blocks
-    dropout: float, default = 0.1
-        Dropout that will be applied internally to the
-        ``TransformerEncoder`` (see
-        :obj:`pytorch_widedeep.model.tab_transformer.TransformerEncoder`) and the
-        output MLP
-    keep_attn_weights: bool, default = False
-        If set to ``True`` the model will store the attention weights in the ``attention_weights``
-        attribute.
-    ff_hidden_dim: int, default = 128
-        Hidden dimension of the ``FeedForward`` Layer. See
-        :obj:`pytorch_widedeep.model.tab_transformer.FeedForward`.
-    transformer_activation: str, default = "gelu"
-        Transformer Encoder activation function
-    with_special_token: bool, default = False,
-        Boolean indicating if the special token ``[CLS]`` was used during the
-        data pre-processing
-    embed_continuous: bool, default = False,
-        Boolean indicating if the continuous features will be "embedded"
-        (i.e. each passed through a 1 layer MLP)
-    cont_norm_layer: str, default =  "layernorm",
-        Type of normalization layer applied to the continuous features if they
-        are not embedded. Options are: 'layernorm' or 'batchnorm'.
-    mlp_hidden_dims: List, Optional, default = None
-        MLP hidden dimensions. If not provided it will default to ``[4*l,
-        2*l]`` where ``l`` is the mlp input dimension
-    mlp_activation: str, default = "gelu"
-        MLP activation function
-    mlp_batchnorm: bool, default = False
-        Boolean indicating whether or not to apply batch normalization to the
-        dense layers
-    mlp_batchnorm_last: bool, default = False
-        Boolean indicating whether or not to apply batch normalization to the
-        last of the dense layers
-    mlp_linear_first: bool, default = False
-        Boolean indicating whether the order of the operations in the dense
-        layer. If ``True: [LIN -> ACT -> BN -> DP]``. If ``False: [BN -> DP ->
-        LIN -> ACT]``
-
-    Attributes
-    ----------
-    cat_embed_layers: ``nn.ModuleDict``
-        Dict with the embeddings per column
-    cont_embed_layers: ``nn.ModuleDict``
-        Dict with the embeddings per column if ``embed_continuous=True``
-    cont_norm_layer: NormLayers
-        Continuous normalization layer if ``continuous_cols`` is not None
-    transformer_blks: ``nn.Sequential``
-        Sequence of Transformer blocks
-    attention_weights: List
-        List with the attention weights per block
-    transformer_mlp: ``nn.Module``
-        MLP component in the model
-    output_dim: int
-        The output dimension of the model. This is a required attribute
-        neccesary to build the WideDeep class
-
-    Example
-    --------
-    >>> import torch
-    >>> from pytorch_widedeep.models import SAINT
-    >>> X_tab = torch.cat((torch.empty(5, 4).random_(4), torch.rand(5, 1)), axis=1)
-    >>> colnames = ['a', 'b', 'c', 'd', 'e']
-    >>> embed_input = [(u,i) for u,i in zip(colnames[:4], [4]*4)]
-    >>> continuous_cols = ['e']
-    >>> column_idx = {k:v for v,k in enumerate(colnames)}
-    >>> model = SAINT(column_idx=column_idx, embed_input=embed_input, continuous_cols=continuous_cols)
-    >>> out = model(X_tab)
-    """
-
-    def __init__(
-        self,
-        column_idx: Dict[str, int],
-        embed_input: List[Tuple[str, int]],
-        continuous_cols: Optional[List[str]] = None,
-        embed_dropout: float = 0.1,
-        full_embed_dropout: bool = False,
-        shared_embed: bool = False,
-        add_shared_embed: bool = False,
-        frac_shared_embed: int = 8,
-        input_dim: int = 32,
-        n_heads: int = 8,
-        n_blocks: int = 6,
-        dropout: float = 0.1,
-        keep_attn_weights: bool = False,
-        ff_hidden_dim: int = 32 * 4,
-        transformer_activation: str = "gelu",
-        with_special_token: bool = False,
-        embed_continuous: bool = False,
-        cont_norm_layer: str = "layernorm",
-        mlp_hidden_dims: Optional[List[int]] = None,
-        mlp_activation: str = "relu",
-        mlp_batchnorm: bool = False,
-        mlp_batchnorm_last: bool = False,
-        mlp_linear_first: bool = True,
-    ):
-        super().__init__(
-            column_idx,
-            embed_input,
-            continuous_cols,
-            embed_dropout,
-            full_embed_dropout,
-            shared_embed,
-            add_shared_embed,
-            frac_shared_embed,
-            input_dim,
-            n_heads,
-            n_blocks,
-            dropout,
-            keep_attn_weights,
-            ff_hidden_dim,
-            transformer_activation,
-            with_special_token,
-            embed_continuous,
-            cont_norm_layer,
-            mlp_hidden_dims,
-            mlp_activation,
-            mlp_batchnorm,
-            mlp_batchnorm_last,
-            mlp_linear_first,
-        )
-
-        if embed_continuous:
-            n_feats = len(embed_input) + len(continuous_cols)
-        else:
-            n_feats = len(embed_input)
-        self.transformer_blks = nn.Sequential()
-        for i in range(n_blocks):
-            self.transformer_blks.add_module(
-                "block" + str(i),
-                SaintEncoder(
-                    input_dim,
-                    n_heads,
-                    keep_attn_weights,
-                    ff_hidden_dim,
-                    dropout,
-                    transformer_activation,
-                    n_feats,
-                ),
-            )
